Failed approach of emulating music's octave law
I devised a complicated method to fit the power of two law
to analyser lines, this means, the frequency width of a musical
note one octave higher is double, for example. I am giving up
and trying something else but before that I keep the failed
approach as a commit.